+++ python/cudf/cudf/_libxx/column.pyx
@@ -18,41 +18,11 @@ from libcpp.memory cimport unique_ptr, make_unique
 
 import cudf._libxx as libcudfxx
 from cudf._libxx.lib cimport *
+from cudf._libxx.lib import *
 from cudf._libxx.null_mask import bitmask_allocation_size_bytes
 
 from cudf.core.buffer import Buffer
 from cudf.utils.dtypes import is_categorical_dtype
-
-
-np_to_cudf_types = {
-    np.dtype('int8'): INT8,
-    np.dtype('int16'): INT16,
-    np.dtype('int32'): INT32,
-    np.dtype('int64'): INT64,
-    np.dtype('float32'): FLOAT32,
-    np.dtype('float64'): FLOAT64,
-    np.dtype("datetime64[s]"): TIMESTAMP_SECONDS,
-    np.dtype("datetime64[ms]"): TIMESTAMP_MILLISECONDS,
-    np.dtype("datetime64[us]"): TIMESTAMP_MICROSECONDS,
-    np.dtype("datetime64[ns]"): TIMESTAMP_NANOSECONDS,
-    np.dtype("object"): STRING,
-    np.dtype("bool"): BOOL8
-}
-
-cudf_to_np_types = {
-    INT8: np.dtype('int8'),
-    INT16: np.dtype('int16'),
-    INT32: np.dtype('int32'),
-    INT64: np.dtype('int64'),
-    FLOAT32: np.dtype('float32'),
-    FLOAT64: np.dtype('float64'),
-    TIMESTAMP_SECONDS: np.dtype("datetime64[s]"),
-    TIMESTAMP_MILLISECONDS: np.dtype("datetime64[ms]"),
-    TIMESTAMP_MICROSECONDS: np.dtype("datetime64[us]"),
-    TIMESTAMP_NANOSECONDS: np.dtype("datetime64[ns]"),
-    STRING: np.dtype("object"),
-    BOOL8: np.dtype("bool")
-}
 
 
 @cython.auto_pickle(True)
@@ -412,7 +382,7 @@ cdef class Column:
         has_nulls = c_col.get()[0].has_nulls()
 
         # After call to release(), c_col is unusable
-        cdef column_contents contents = c_col.get()[0].release()
+        cdef column_contents contents = move(c_col.get()[0].release())
 
         data = DeviceBuffer.c_from_unique_ptr(move(contents.data))
         data = Buffer(data)