+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - PR #4164 Add Buffer "constructor-kwargs" header
 - PR #4159 Add COUNT aggregation that includes null values
 - PR #4190 Add libcudf++ transpose Cython implementation
+- PR #4216 Add cudf.Scalar Python type
 
 ## Improvements
 
@@ -144,6 +145,7 @@
 - PR #4201 Fix java window tests
 - PR #4199 Fix potential race condition in memcpy_block
 - PR #4221 Fix series dict alignment to not drop index name
+- PR #4218 Fix `get_aggregation` definition with `except *`
 - PR #4215 Fix performance regression in strings::detail::concatenate
 - PR #4214 Alter ValueError exception for GPU accelerated Parquet writer to properly report `categorical` columns are not supported.